requirements.txt
opencv-python
ultralytics

google-genai
python-dotenv
httpx